Remove .themes segment from int-ui packages
