+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/IntUiBridge.kt
@@ -25,66 +25,66 @@ import org.jetbrains.jewel.CheckboxState
 import org.jetbrains.jewel.IntelliJComponentStyling
 import org.jetbrains.jewel.IntelliJThemeIconData
 import org.jetbrains.jewel.SvgLoader
+import org.jetbrains.jewel.intui.core.IntUiThemeDefinition
+import org.jetbrains.jewel.intui.standalone.styling.IntUiButtonColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiButtonMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiButtonStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiCheckboxColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiCheckboxIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiCheckboxMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiCheckboxStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiChipColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiChipMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiChipStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiDropdownColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiDropdownIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiDropdownMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiDropdownStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiGroupHeaderColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiGroupHeaderMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiGroupHeaderStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiHorizontalProgressBarColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiHorizontalProgressBarMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiHorizontalProgressBarStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLabelledTextFieldColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLabelledTextFieldMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLabelledTextFieldStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLabelledTextFieldTextStyles
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLazyTreeColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLazyTreeIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLazyTreeMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLazyTreeStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLinkColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLinkIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLinkMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLinkStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiLinkTextStyles
+import org.jetbrains.jewel.intui.standalone.styling.IntUiMenuColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiMenuIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiMenuItemColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiMenuItemMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiMenuMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiMenuStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiRadioButtonColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiRadioButtonIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiRadioButtonMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiRadioButtonStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiScrollbarColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiScrollbarMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiScrollbarStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiSubmenuMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTabColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTabContentAlpha
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTabIcons
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTabMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTabStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTextAreaColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTextAreaMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTextAreaStyle
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTextFieldColors
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTextFieldMetrics
+import org.jetbrains.jewel.intui.standalone.styling.IntUiTextFieldStyle
 import org.jetbrains.jewel.styling.InputFieldStyle
-import org.jetbrains.jewel.themes.intui.core.IntUiThemeDefinition
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiButtonColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiButtonMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiButtonStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiCheckboxColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiCheckboxIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiCheckboxMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiCheckboxStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiChipColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiChipMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiChipStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiDropdownColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiDropdownIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiDropdownMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiDropdownStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiGroupHeaderColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiGroupHeaderMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiGroupHeaderStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiHorizontalProgressBarColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiHorizontalProgressBarMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiHorizontalProgressBarStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLabelledTextFieldColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLabelledTextFieldMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLabelledTextFieldStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLabelledTextFieldTextStyles
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLazyTreeColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLazyTreeIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLazyTreeMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLazyTreeStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLinkColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLinkIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLinkMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLinkStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiLinkTextStyles
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiMenuColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiMenuIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiMenuItemColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiMenuItemMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiMenuMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiMenuStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiRadioButtonColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiRadioButtonIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiRadioButtonMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiRadioButtonStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiScrollbarColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiScrollbarMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiScrollbarStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiSubmenuMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTabColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTabContentAlpha
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTabIcons
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTabMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTabStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTextAreaColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTextAreaMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTextAreaStyle
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTextFieldColors
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTextFieldMetrics
-import org.jetbrains.jewel.themes.intui.standalone.styling.IntUiTextFieldStyle
 import org.jetbrains.skiko.DependsOnJBR
 import javax.swing.UIManager
 import kotlin.time.Duration.Companion.milliseconds
