+++ int-ui/int-ui-standalone/src/main/kotlin/org/jetbrains/jewel/intui/standalone/styling/IntUiTextAreaStyling.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.jewel.themes.intui.standalone.styling
+package org.jetbrains.jewel.intui.standalone.styling
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.shape.CornerSize
@@ -10,54 +10,40 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import org.jetbrains.jewel.DropdownState
-import org.jetbrains.jewel.SvgLoader
-import org.jetbrains.jewel.styling.DropdownColors
-import org.jetbrains.jewel.styling.DropdownIcons
-import org.jetbrains.jewel.styling.DropdownMetrics
-import org.jetbrains.jewel.styling.DropdownStyle
-import org.jetbrains.jewel.styling.MenuStyle
-import org.jetbrains.jewel.styling.PainterProvider
-import org.jetbrains.jewel.styling.ResourcePainterProvider
-import org.jetbrains.jewel.themes.intui.core.theme.IntUiDarkTheme
-import org.jetbrains.jewel.themes.intui.core.theme.IntUiLightTheme
-import org.jetbrains.jewel.themes.intui.standalone.IntUiTheme
+import org.jetbrains.jewel.intui.core.theme.IntUiDarkTheme
+import org.jetbrains.jewel.intui.core.theme.IntUiLightTheme
+import org.jetbrains.jewel.intui.standalone.IntUiTheme
+import org.jetbrains.jewel.styling.InputFieldMetrics
+import org.jetbrains.jewel.styling.TextAreaColors
+import org.jetbrains.jewel.styling.TextAreaStyle
 
 @Stable
-data class IntUiDropdownStyle(
-    override val colors: IntUiDropdownColors,
-    override val metrics: IntUiDropdownMetrics,
-    override val icons: IntUiDropdownIcons,
+data class IntUiTextAreaStyle(
+    override val colors: IntUiTextAreaColors,
+    override val metrics: IntUiTextAreaMetrics,
     override val textStyle: TextStyle,
-    override val menuStyle: MenuStyle,
-) : DropdownStyle {
+) : TextAreaStyle {
 
     companion object {
 
         @Composable
         fun light(
-            svgLoader: SvgLoader,
-            colors: IntUiDropdownColors = IntUiDropdownColors.light(),
-            metrics: IntUiDropdownMetrics = IntUiDropdownMetrics(),
-            icons: IntUiDropdownIcons = intUiDropdownIcons(svgLoader),
+            colors: IntUiTextAreaColors = IntUiTextAreaColors.light(),
+            metrics: IntUiTextAreaMetrics = IntUiTextAreaMetrics(),
             textStyle: TextStyle = IntUiTheme.defaultLightTextStyle,
-            menuStyle: MenuStyle = IntUiMenuStyle.light(svgLoader),
-        ) = IntUiDropdownStyle(colors, metrics, icons, textStyle, menuStyle)
+        ) = IntUiTextAreaStyle(colors, metrics, textStyle)
 
         @Composable
         fun dark(
-            svgLoader: SvgLoader,
-            colors: IntUiDropdownColors = IntUiDropdownColors.dark(),
-            metrics: IntUiDropdownMetrics = IntUiDropdownMetrics(),
-            icons: IntUiDropdownIcons = intUiDropdownIcons(svgLoader),
+            colors: IntUiTextAreaColors = IntUiTextAreaColors.dark(),
+            metrics: IntUiTextAreaMetrics = IntUiTextAreaMetrics(),
             textStyle: TextStyle = IntUiTheme.defaultDarkTextStyle,
-            menuStyle: MenuStyle = IntUiMenuStyle.dark(svgLoader),
-        ) = IntUiDropdownStyle(colors, metrics, icons, textStyle, menuStyle)
+        ) = IntUiTextAreaStyle(colors, metrics, textStyle)
     }
 }
 
 @Immutable
-data class IntUiDropdownColors(
+data class IntUiTextAreaColors(
     override val background: Color,
     override val backgroundDisabled: Color,
     override val backgroundFocused: Color,
@@ -73,12 +59,13 @@ data class IntUiDropdownColors(
     override val borderFocused: Color,
     override val borderPressed: Color,
     override val borderHovered: Color,
-    override val iconTint: Color,
-    override val iconTintDisabled: Color,
-    override val iconTintFocused: Color,
-    override val iconTintPressed: Color,
-    override val iconTintHovered: Color,
-) : DropdownColors {
+    override val caret: Color,
+    override val caretDisabled: Color,
+    override val caretFocused: Color,
+    override val caretPressed: Color,
+    override val caretHovered: Color,
+    override val placeholder: Color,
+) : TextAreaColors {
 
     companion object {
 
@@ -99,12 +86,13 @@ data class IntUiDropdownColors(
             borderFocused: Color = IntUiLightTheme.colors.blue(4),
             borderPressed: Color = border,
             borderHovered: Color = border,
-            iconTint: Color = IntUiLightTheme.colors.grey(7),
-            iconTintDisabled: Color = IntUiLightTheme.colors.grey(9),
-            iconTintFocused: Color = iconTint,
-            iconTintPressed: Color = iconTint,
-            iconTintHovered: Color = iconTint,
-        ) = IntUiDropdownColors(
+            caret: Color = IntUiLightTheme.colors.grey(1),
+            caretDisabled: Color = caret,
+            caretFocused: Color = caret,
+            caretPressed: Color = caret,
+            caretHovered: Color = caret,
+            placeholder: Color = IntUiLightTheme.colors.grey(8),
+        ) = IntUiTextAreaColors(
             background,
             backgroundDisabled,
             backgroundFocused,
@@ -120,11 +108,12 @@ data class IntUiDropdownColors(
             borderFocused,
             borderPressed,
             borderHovered,
-            iconTint,
-            iconTintDisabled,
-            iconTintFocused,
-            iconTintPressed,
-            iconTintHovered,
+            caret,
+            caretDisabled,
+            caretFocused,
+            caretPressed,
+            caretHovered,
+            placeholder,
         )
 
         @Composable
@@ -140,16 +129,17 @@ data class IntUiDropdownColors(
             contentPressed: Color = content,
             contentHovered: Color = content,
             border: Color = IntUiDarkTheme.colors.grey(5),
-            borderDisabled: Color = IntUiDarkTheme.colors.grey(11),
+            borderDisabled: Color = border,
             borderFocused: Color = IntUiDarkTheme.colors.blue(6),
             borderPressed: Color = border,
             borderHovered: Color = border,
-            iconTint: Color = IntUiDarkTheme.colors.grey(10),
-            iconTintDisabled: Color = IntUiDarkTheme.colors.grey(6),
-            iconTintFocused: Color = iconTint,
-            iconTintPressed: Color = iconTint,
-            iconTintHovered: Color = iconTint,
-        ) = IntUiDropdownColors(
+            caret: Color = IntUiDarkTheme.colors.grey(12),
+            caretDisabled: Color = caret,
+            caretFocused: Color = caret,
+            caretPressed: Color = caret,
+            caretHovered: Color = caret,
+            placeholder: Color = IntUiDarkTheme.colors.grey(7),
+        ) = IntUiTextAreaColors(
             background,
             backgroundDisabled,
             backgroundFocused,
@@ -165,43 +155,20 @@ data class IntUiDropdownColors(
             borderFocused,
             borderPressed,
             borderHovered,
-            iconTint,
-            iconTintDisabled,
-            iconTintFocused,
-            iconTintPressed,
-            iconTintHovered,
+            caret,
+            caretDisabled,
+            caretFocused,
+            caretPressed,
+            caretHovered,
+            placeholder,
         )
     }
 }
 
 @Stable
-data class IntUiDropdownMetrics(
-    override val arrowMinSize: DpSize = DpSize((23 + 3).dp, 24.dp),
-    override val minSize: DpSize = DpSize((49 + 23 + 6).dp, 24.dp),
+data class IntUiTextAreaMetrics(
     override val cornerSize: CornerSize = CornerSize(4.dp),
-    override val contentPadding: PaddingValues = PaddingValues(horizontal = 6.dp, vertical = 3.dp),
+    override val contentPadding: PaddingValues = PaddingValues(horizontal = 6.dp, vertical = 2.dp),
+    override val minSize: DpSize = DpSize(144.dp, 28.dp),
     override val borderWidth: Dp = 1.dp,
-) : DropdownMetrics
-
-@Immutable
-data class IntUiDropdownIcons(
-    override val chevronDown: PainterProvider<DropdownState>,
-) : DropdownIcons {
-
-    companion object {
-
-        @Composable
-        fun chevronDown(
-            svgLoader: SvgLoader,
-            basePath: String = "icons/intui/chevronDown.svg",
-        ): PainterProvider<DropdownState> =
-            ResourcePainterProvider.stateful(basePath, svgLoader)
-    }
-}
-
-@Composable
-fun intUiDropdownIcons(
-    svgLoader: SvgLoader,
-    chevronDown: PainterProvider<DropdownState> = IntUiDropdownIcons.chevronDown(svgLoader),
-) =
-    IntUiDropdownIcons(chevronDown)
+) : InputFieldMetrics
