+++ int-ui/int-ui-standalone/src/main/kotlin/org/jetbrains/jewel/intui/standalone/styling/IntUiTextFieldStyling.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.jewel.themes.intui.standalone.styling
+package org.jetbrains.jewel.intui.standalone.styling
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.shape.CornerSize
@@ -10,45 +10,40 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import org.jetbrains.jewel.styling.LabelledTextFieldColors
-import org.jetbrains.jewel.styling.LabelledTextFieldMetrics
-import org.jetbrains.jewel.styling.LabelledTextFieldStyle
-import org.jetbrains.jewel.styling.LabelledTextFieldTextStyles
-import org.jetbrains.jewel.themes.intui.core.theme.IntUiDarkTheme
-import org.jetbrains.jewel.themes.intui.core.theme.IntUiLightTheme
-import org.jetbrains.jewel.themes.intui.standalone.IntUiTheme
+import org.jetbrains.jewel.intui.core.theme.IntUiDarkTheme
+import org.jetbrains.jewel.intui.core.theme.IntUiLightTheme
+import org.jetbrains.jewel.intui.standalone.IntUiTheme
+import org.jetbrains.jewel.styling.InputFieldMetrics
+import org.jetbrains.jewel.styling.TextFieldColors
+import org.jetbrains.jewel.styling.TextFieldStyle
 
 @Stable
-data class IntUiLabelledTextFieldStyle(
-    override val colors: IntUiLabelledTextFieldColors,
-    override val metrics: IntUiLabelledTextFieldMetrics,
+data class IntUiTextFieldStyle(
+    override val colors: IntUiTextFieldColors,
+    override val metrics: IntUiTextFieldMetrics,
     override val textStyle: TextStyle,
-    override val textStyles: IntUiLabelledTextFieldTextStyles,
-) : LabelledTextFieldStyle {
+) : TextFieldStyle {
 
     companion object {
 
         @Composable
         fun light(
-            colors: IntUiLabelledTextFieldColors = IntUiLabelledTextFieldColors.light(),
-            metrics: IntUiLabelledTextFieldMetrics = IntUiLabelledTextFieldMetrics(),
+            colors: IntUiTextFieldColors = IntUiTextFieldColors.light(),
+            metrics: IntUiTextFieldMetrics = IntUiTextFieldMetrics(),
             textStyle: TextStyle = IntUiTheme.defaultLightTextStyle,
-            textStyles: IntUiLabelledTextFieldTextStyles = IntUiLabelledTextFieldTextStyles.light(),
-        ) = IntUiLabelledTextFieldStyle(colors, metrics, textStyle, textStyles)
+        ) = IntUiTextFieldStyle(colors, metrics, textStyle)
 
         @Composable
         fun dark(
-            colors: IntUiLabelledTextFieldColors = IntUiLabelledTextFieldColors.dark(),
-            metrics: IntUiLabelledTextFieldMetrics = IntUiLabelledTextFieldMetrics(),
+            colors: IntUiTextFieldColors = IntUiTextFieldColors.dark(),
+            metrics: IntUiTextFieldMetrics = IntUiTextFieldMetrics(),
             textStyle: TextStyle = IntUiTheme.defaultDarkTextStyle,
-            textStyles: IntUiLabelledTextFieldTextStyles = IntUiLabelledTextFieldTextStyles.dark(),
-        ) = IntUiLabelledTextFieldStyle(colors, metrics, textStyle, textStyles)
+        ) = IntUiTextFieldStyle(colors, metrics, textStyle)
     }
 }
 
 @Immutable
-data class IntUiLabelledTextFieldColors(
+data class IntUiTextFieldColors(
     override val background: Color,
     override val backgroundDisabled: Color,
     override val backgroundFocused: Color,
@@ -70,9 +65,7 @@ data class IntUiLabelledTextFieldColors(
     override val caretPressed: Color,
     override val caretHovered: Color,
     override val placeholder: Color,
-    override val label: Color,
-    override val hint: Color,
-) : LabelledTextFieldColors {
+) : TextFieldColors {
 
     companion object {
 
@@ -99,9 +92,7 @@ data class IntUiLabelledTextFieldColors(
             caretPressed: Color = caret,
             caretHovered: Color = caret,
             placeholder: Color = IntUiLightTheme.colors.grey(8),
-            label: Color = IntUiLightTheme.colors.grey(1),
-            hint: Color = IntUiLightTheme.colors.grey(6),
-        ) = IntUiLabelledTextFieldColors(
+        ) = IntUiTextFieldColors(
             background,
             backgroundDisabled,
             backgroundFocused,
@@ -123,8 +114,6 @@ data class IntUiLabelledTextFieldColors(
             caretPressed,
             caretHovered,
             placeholder,
-            label,
-            hint,
         )
 
         @Composable
@@ -150,9 +139,7 @@ data class IntUiLabelledTextFieldColors(
             caretPressed: Color = caret,
             caretHovered: Color = caret,
             placeholder: Color = IntUiDarkTheme.colors.grey(7),
-            label: Color = IntUiDarkTheme.colors.grey(12),
-            hint: Color = IntUiDarkTheme.colors.grey(7),
-        ) = IntUiLabelledTextFieldColors(
+        ) = IntUiTextFieldColors(
             background,
             backgroundDisabled,
             backgroundFocused,
@@ -174,46 +161,14 @@ data class IntUiLabelledTextFieldColors(
             caretPressed,
             caretHovered,
             placeholder,
-            label,
-            hint,
         )
     }
 }
 
 @Stable
-data class IntUiLabelledTextFieldMetrics(
+data class IntUiTextFieldMetrics(
     override val cornerSize: CornerSize = CornerSize(4.dp),
     override val contentPadding: PaddingValues = PaddingValues(horizontal = 9.dp, vertical = 6.dp),
-    override val minSize: DpSize = DpSize(49.dp, 24.dp),
+    override val minSize: DpSize = DpSize(144.dp, 28.dp),
     override val borderWidth: Dp = 1.dp,
-    override val labelSpacing: Dp = 6.dp,
-    override val hintSpacing: Dp = 6.dp,
-) : LabelledTextFieldMetrics
-
-@Immutable
-data class IntUiLabelledTextFieldTextStyles(
-    override val label: TextStyle,
-    override val hint: TextStyle,
-) : LabelledTextFieldTextStyles {
-
-    companion object {
-
-        @Composable
-        fun light(
-            label: TextStyle = IntUiTheme.defaultLightTextStyle,
-            hint: TextStyle = IntUiTheme.defaultLightTextStyle.copy(
-                fontSize = 12.sp,
-                lineHeight = 16.sp,
-            ),
-        ) = IntUiLabelledTextFieldTextStyles(label, hint)
-
-        @Composable
-        fun dark(
-            label: TextStyle = IntUiTheme.defaultDarkTextStyle,
-            hint: TextStyle = IntUiTheme.defaultDarkTextStyle.copy(
-                fontSize = 12.sp,
-                lineHeight = 16.sp,
-            ),
-        ) = IntUiLabelledTextFieldTextStyles(label, hint)
-    }
-}
+) : InputFieldMetrics
